+++ backend/eslint.config.mjs
@@ -25,26 +25,6 @@ export default tseslint.config(
       // New in ESLint 10 recommended - existing patterns, suppress until addressed
       'no-useless-assignment': 'warn',
       'preserve-caught-error': 'warn',
-      // The private @studio-saelix/sencho-pro package is loaded at
-      // runtime via dynamic import in entitlements/loadProvider.ts.
-      // A static import would (a) bundle the package into the public
-      // BSL build via TypeScript's module resolution, defeating the
-      // privacy split, and (b) break in Community-only environments
-      // where the package is not installed. Use loadEntitlementProvider()
-      // and the registry instead.
-      // no-restricted-imports flags ES `import` statements only by
-      // default; dynamic `import()` expressions are not flagged unless
-      // we set `allowDynamicImports: false`. The loader uses
-      // `await import('@studio-saelix/sencho-pro')` and is therefore
-      // not affected without an extra opt-in.
-      'no-restricted-imports': ['error', {
-        patterns: [
-          {
-            group: ['@studio-saelix/sencho-pro', '@studio-saelix/sencho-pro/*'],
-            message: 'Do not statically import the private package. Use loadEntitlementProvider() and getEntitlementProvider() from src/entitlements/.',
-          },
-        ],
-      }],
     },
   },
 )
