+++ backend/src/bootstrap/startup.ts
@@ -1,8 +1,7 @@
 import type { Server } from 'http';
 import { FileSystemService } from '../services/FileSystemService';
 import { NodeRegistry } from '../services/NodeRegistry';
-import { loadEntitlementProvider } from '../entitlements/loadProvider';
-import { setEntitlementProvider } from '../entitlements/registry';
+import { LicenseService } from '../services/LicenseService';
 import SelfUpdateService from '../services/SelfUpdateService';
 import { MonitorService } from '../services/MonitorService';
 import { AutoHealService } from '../services/AutoHealService';
@@ -32,14 +31,8 @@ export async function startServer(server: Server): Promise<void> {
     console.error('Migration failed:', error);
   }
 
-  // Resolve the EntitlementProvider before any tier-gated code can run.
-  // Phase 1 returns the in-tree LicenseService singleton; Phase 2 will
-  // dynamic-import the private package and fall back to Community.
-  // Awaited because the loader is async-by-signature (so the Phase-2
-  // swap doesn't change this call site).
-  const entitlementProvider = await loadEntitlementProvider();
-  setEntitlementProvider(entitlementProvider);
-  entitlementProvider.initialize();
+  // Initialize the license service before any tier-gated code can run.
+  LicenseService.getInstance().initialize();
 
   // Synchronous starts: schedule background timers and continue. None of
   // these fire their first tick for at least a few seconds, so they
