+++ backend/src/websocket/hostConsole.ts
@@ -5,14 +5,14 @@ import path from 'path';
 import { FileSystemService } from '../services/FileSystemService';
 import { NodeRegistry } from '../services/NodeRegistry';
 import { HostTerminalService } from '../services/HostTerminalService';
-import { PROXY_TIER_HEADER, PROXY_VARIANT_HEADER } from '../entitlements/headers';
+import { PROXY_TIER_HEADER, PROXY_VARIANT_HEADER } from '../services/license-headers';
 import {
   isLicenseTier,
   isLicenseVariant,
   normalizeTier,
   normalizeVariant,
-} from '../entitlements/normalize';
-import { getEntitlementProvider } from '../entitlements/registry';
+} from '../services/license-normalize';
+import { LicenseService } from '../services/LicenseService';
 import { ROLE_PERMISSIONS, type PermissionAction } from '../middleware/permissions';
 import type { UserRole } from '../services/DatabaseService';
 import { getErrorMessage } from '../utils/errors';
@@ -64,7 +64,7 @@ export function handleHostConsoleWs(
 
   const consoleTierHeader = req.headers[PROXY_TIER_HEADER] as string | undefined;
   const consoleVariantHeader = req.headers[PROXY_VARIANT_HEADER] as string | undefined;
-  const ls = getEntitlementProvider();
+  const ls = LicenseService.getInstance();
   const consoleTier = (isConsoleSession && isLicenseTier(consoleTierHeader))
     ? normalizeTier(consoleTierHeader)
     : ls.getTier();
